# src/main/resources/openapi/triage-api.yaml
openapi: 3.0.0
info:
  title: API Sistema de Triage Académico
  version: 1.0.0
  description: API para gestionar solicitudes académicas y su ciclo de vida.

servers:
  - url: http://localhost:8080

paths:

  /api/solicitudes:
    post:
      summary: Registrar una nueva solicitud
      operationId: registrarSolicitud
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SolicitudCreateRequest'
      responses:
        '201':
          description: Solicitud creada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudResponse'

    get:
      summary: Listar solicitudes con filtros
      operationId: listarSolicitudes
      parameters:
        - name: estado
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/EstadoSolicitud'
        - name: tipo
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/TipoSolicitud'
        - name: prioridad
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Prioridad'
        - name: responsableId
          in: query
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Lista de solicitudes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SolicitudResponse'

  /api/solicitudes/{id}:
    get:
      summary: Obtener solicitud por id
      operationId: obtenerSolicitudPorId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Solicitud encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudResponse'
        '404':
          description: Solicitud no encontrada

  /api/solicitudes/{id}/clasificar:
    patch:
      summary: Clasificar una solicitud
      operationId: clasificarSolicitud
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: usuarioId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClasificarSolicitudRequest'
      responses:
        '200':
          description: Solicitud clasificada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudResponse'
        '409':
          description: Transición inválida
        '403':
          description: Usuario no autorizado

  /api/solicitudes/{id}/prioridad:
    put:
      summary: Asignar prioridad a una solicitud
      operationId: asignarPrioridad
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: usuarioId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AsignarPrioridadRequest'
      responses:
        '200':
          description: Prioridad asignada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudResponse'
        '403':
          description: Usuario no autorizado

  /api/solicitudes/{id}/asignar:
    patch:
      summary: Asignar responsable a una solicitud
      operationId: asignarResponsable
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: usuarioId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AsignarResponsableRequest'
      responses:
        '200':
          description: Responsable asignado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudResponse'
        '403':
          description: Usuario no autorizado

  /api/solicitudes/{id}/historial:
    get:
      summary: Obtener historial de una solicitud
      operationId: obtenerHistorial
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Historial encontrado
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HistorialSolicitudResponse'

  /api/solicitudes/{id}/estado:
    patch:
      summary: Cambiar el estado de una solicitud
      operationId: cambiarEstado
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CambiarEstadoRequest'
      responses:
        '200':
          description: Estado cambiado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudResponse'
        '409':
          description: Transición inválida

  /api/solicitudes/{id}/cerrar:
    put:
      summary: Cerrar una solicitud
      operationId: cerrarSolicitud
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: usuarioId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CerrarSolicitudRequest'
      responses:
        '200':
          description: Solicitud cerrada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SolicitudResponse'
        '403':
          description: Usuario no autorizado

  /api/solicitudes/{id}/resumen:
    get:
      summary: Generar resumen de una solicitud
      operationId: generarResumen
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Resumen generado correctamente
          content:
            text/plain:
              schema:
                type: string

  /api/solicitudes/sugerir-clasificacion:
    post:
      summary: Generar sugerencia automática de clasificación
      operationId: sugerirClasificacion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SugerirClasificacionRequest'
      responses:
        '200':
          description: Sugerencia generada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SugerenciaClasificacionResponse'

components:
  schemas:

    SolicitudResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        tipo:
          $ref: '#/components/schemas/TipoSolicitud'
        descripcion:
          type: string
        canalOrigen:
          $ref: '#/components/schemas/CanalOrigen'
        fechaRegistro:
          type: string
          format: date-time
        solicitanteId:
          type: integer
          format: int64
        responsableAsignadoId:
          type: integer
          format: int64
          nullable: true
        estado:
          $ref: '#/components/schemas/EstadoSolicitud'
        prioridad:
          $ref: '#/components/schemas/Prioridad'
        justificacionPrioridad:
          type: string
          nullable: true
        impactoAcademico:
          $ref: '#/components/schemas/ImpactoAcademico'
        fechaLimite:
          type: string
          format: date
          nullable: true
        observacionCierre:
          type: string
          nullable: true
        historial:
          type: array
          items:
            $ref: '#/components/schemas/HistorialSolicitudResponse'

    HistorialSolicitudResponse:
      type: object
      properties:
        fechaHora:
          type: string
          format: date-time
        accion:
          type: string
        observacion:
          type: string
        usuarioResponsableId:
          type: integer
          format: int64

    SolicitudCreateRequest:
      type: object
      required:
        - tipo
        - descripcion
        - canalOrigen
        - solicitanteId
      properties:
        tipo:
          $ref: '#/components/schemas/TipoSolicitud'
        descripcion:
          type: string
        canalOrigen:
          $ref: '#/components/schemas/CanalOrigen'
        solicitanteId:
          type: integer
          format: int64

    ClasificarSolicitudRequest:
      type: object
      required:
        - tipo
      properties:
        tipo:
          $ref: '#/components/schemas/TipoSolicitud'

    AsignarPrioridadRequest:
      type: object
      required:
        - impactoAcademico
        - fechaLimite
      properties:
        impactoAcademico:
          $ref: '#/components/schemas/ImpactoAcademico'
        fechaLimite:
          type: string
          format: date

    AsignarResponsableRequest:
      type: object
      required:
        - responsableId
      properties:
        responsableId:
          type: integer
          format: int64

    CambiarEstadoRequest:
      type: object
      required:
        - nuevoEstado
        - usuarioId
      properties:
        nuevoEstado:
          $ref: '#/components/schemas/EstadoSolicitud'
        observacion:
          type: string
        usuarioId:
          type: integer
          format: int64

    CerrarSolicitudRequest:
      type: object
      required:
        - observacionCierre
      properties:
        observacionCierre:
          type: string

    SugerirClasificacionRequest:
      type: object
      required:
        - descripcion
      properties:
        descripcion:
          type: string
          minLength: 10
          maxLength: 1000

    SugerenciaClasificacionResponse:
      type: object
      properties:
        tipoSugerido:
          $ref: '#/components/schemas/TipoSolicitud'
        prioridadSugerida:
          $ref: '#/components/schemas/Prioridad'
        justificacion:
          type: string

    EstadoSolicitud:
      type: string
      enum:
        - REGISTRADA
        - CLASIFICADA
        - EN_ATENCION
        - ATENDIDA
        - CERRADA

    TipoSolicitud:
      type: string
      enum:
        - REGISTRO_ASIGNATURAS
        - HOMOLOGACION
        - CANCELACION_ASIGNATURAS
        - SOLICITUD_CUPOS
        - CONSULTA_ACADEMICA

    Prioridad:
      type: string
      enum:
        - BAJA
        - MEDIA
        - ALTA
        - CRITICA

    ImpactoAcademico:
      type: string
      enum:
        - BAJO
        - MEDIO
        - ALTO

    CanalOrigen:
      type: string
      enum:
        - PRESENCIAL
        - CORREO
        - SAC
        - CSU
        - TELEFONICO